move css files to styles directory
